:recycle: Remove Query structure
